MyAnimeArchive-server/celery_workers/requirements.txt
celery[redis]
mysql-connector-python==8.0.28
python-decouple==3.5